+++ prism/tests/integration/test_targets.py
@@ -12,21 +12,11 @@ Table of Contents:
 ###########
 
 # Standard library imports
-import io
-import boto3
-import boto3.session
 import os
 from pathlib import Path
-import re
-import shutil
-import unittest
 import pandas as pd
-import yaml
 
 # Prism imports
-import prism.cli.base
-from prism.main import main
-import prism.prism_logging
 import prism.tests.integration.integration_test_class as integration_test_class
 
 
@@ -39,13 +29,12 @@ TEST_CASE_WKDIR = os.path.dirname(__file__)
 TEST_PROJECTS = Path(TEST_CASE_WKDIR) / 'test_projects'
 
 
-################################
-## Test case class definition ##
-################################
+##############################
+# Test case class definition #
+##############################
 
-class TestTargetIntegration(integration_test_class.IntegrationTestCase): 
+class TestTargetIntegration(integration_test_class.IntegrationTestCase):
 
-    
     def test_targets(self):
         """
         `prism spark-submit` on project `008_targets` produces the expected targets
@@ -65,7 +54,7 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
 
         # Execute command.
         args = ['spark-submit']
-        spark_submit = self._run_prism(args)
+        _ = self._run_prism(args)
         self.assertTrue(Path(wkdir / '.compiled').is_dir())
         self.assertTrue(Path(wkdir / '.compiled' / 'manifest.json').is_file())
         manifest = self._load_manifest(Path(wkdir / '.compiled' / 'manifest.json'))
@@ -78,9 +67,9 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
         # Parquet target
         target_parquet_df = pd.read_parquet(Path(wkdir / 'output' / 'target_parquet'))
         expected_df = pd.DataFrame({
-            'col1': ['col1_value1', 'col1_value2', 'col1_value3', 'col1_value4', 'col1_value5', 'col1_value6'],
-            'col2': ['col2_value1', 'col2_value2', 'col2_value3', 'col2_value4', 'col2_value5', 'col2_value6'],
-            'col3': ['col3_value1', 'col3_value2', 'col3_value3', 'col3_value4', 'col3_value5', 'col3_value6'],
+            'col1': ['col1_value1', 'col1_value2', 'col1_value3', 'col1_value4', 'col1_value5', 'col1_value6'],  # noqa: E501
+            'col2': ['col2_value1', 'col2_value2', 'col2_value3', 'col2_value4', 'col2_value5', 'col2_value6'],  # noqa: E501
+            'col3': ['col3_value1', 'col3_value2', 'col3_value3', 'col3_value4', 'col3_value5', 'col3_value6'],  # noqa: E501
         })
         self.assertTrue(target_parquet_df.equals(expected_df))
 
@@ -89,7 +78,9 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
         self.assertEqual('Hello, world!', target_txt_txt)
 
         # CSV (single target)
-        target_csv_df = pd.read_csv(Path(wkdir / 'output' / 'target_csv.csv'), index_col=None)
+        target_csv_df = pd.read_csv(
+            Path(wkdir / 'output' / 'target_csv.csv'), index_col=None
+        )
         expected_df = pd.DataFrame({
             'col1': ['col1_value1', 'col1_value2', 'col1_value3'],
             'col2': ['col2_value1', 'col2_value2', 'col2_value3'],
@@ -101,8 +92,12 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
         self.assertTrue(target_csv_df.equals(expected_df))
 
         # CSV (multiple targets)
-        target_csv_mult_df1 = pd.read_csv(Path(wkdir / 'output' / 'target_csv_mult_df1.csv'), index_col=None)
-        target_csv_mult_df2 = pd.read_csv(Path(wkdir / 'output' / 'target_csv_mult_df2.csv'), index_col=None)
+        target_csv_mult_df1 = pd.read_csv(
+            Path(wkdir / 'output' / 'target_csv_mult_df1.csv'), index_col=None
+        )
+        target_csv_mult_df2 = pd.read_csv(
+            Path(wkdir / 'output' / 'target_csv_mult_df2.csv'), index_col=None
+        )
         expected_df1 = pd.DataFrame({
             'col1': ['col1_value1', 'col1_value2', 'col1_value3'],
             'col2': ['col2_value1', 'col2_value2', 'col2_value3'],
@@ -124,12 +119,17 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
         self.assertTrue(target_csv_mult_df2.equals(expected_df2))
 
         # CSV (iterated targets)
-        target_csv_iter_df1 = pd.read_csv(Path(wkdir / 'output' / 'target_csv_iter_df1.csv'), index_col=None)
-        target_csv_iter_df2 = pd.read_csv(Path(wkdir / 'output' / 'target_csv_iter_df2.csv'), index_col=None)
+        target_csv_iter_df1 = pd.read_csv(
+            Path(wkdir / 'output' / 'target_csv_iter_df1.csv'), index_col=None
+        )
+        target_csv_iter_df2 = pd.read_csv(
+            Path(wkdir / 'output' / 'target_csv_iter_df2.csv'), index_col=None
+        )
         self.assertTrue(target_csv_iter_df1.equals(expected_df1))
         self.assertTrue(target_csv_iter_df2.equals(expected_df2))
 
-        # Remove all files in the output directory (to avoid having to commit and re-commit to Github)
+        # Remove all files in the output directory (to avoid having to commit and
+        # re-commit to Github)
         self._remove_parquet_files_in_dir(Path(wkdir / 'output' / 'target_parquet'))
 
         # Remove all compiled modules
@@ -137,6 +137,3 @@ class TestTargetIntegration(integration_test_class.IntegrationTestCase):
 
         # Set up wkdir for next test
         self._set_up_wkdir()
-
-
-# EOF
