+++ setup.cfg
@@ -73,7 +73,7 @@ testing =
 
 [options.entry_points]
 console_scripts = 
-    prism = prism.main2:invoke
+    prism = prism.main:invoke
 
 
 [flake8]
