--- src/prism
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-
-# Imports
-import os
-import sys
-import re
-from prism.main import main
-
-# Main method
-if __name__ == '__main__':
-    main()
-
-
-# EOF
